requirements.txt
nicegui
flask
flask-cors
pywebview
pyinstaller
pytubefix
# Используем версию из git, так как на PyPI устаревшая
git+https://github.com/m-bain/whisperx.git
certifi
pyannote.audio>=3.1.0
torch
torchaudio
python-dotenv
requests
deep-translator
# NumPy: WhisperX требует numpy>=2.0.2
# После установки numpy 2.x нужно переустановить scipy и sklearn:
# pip install --force-reinstall scipy scikit-learn
numpy>=2.0.2,<2.1.0
# Voice Cloning (Coqui XTTS v2)
# Примечание: TTS требует Python 3.10+
# Если Python < 3.10, используйте: brew install python@3.11
# Затем создайте venv: python3.11 -m venv venv && source venv/bin/activate
# Для PyTorch 2.9+ требуется: coqui-tts[codec]
coqui-tts[codec]>=0.27.0
pydub
# Video Assembly
moviepy>=1.0.3